Rename unsafe Java test suites to match Scala test naming convention.
This is necessary in order for SBT to recognize and run these suites.
