+++ unsafe/src/test/java/org/apache/spark/unsafe/array/LongArraySuite.java
@@ -15,15 +15,24 @@
  * limitations under the License.
  */
 
-package org.apache.spark.unsafe.map;
+package org.apache.spark.unsafe.array;
 
-import org.apache.spark.unsafe.memory.MemoryAllocator;
+import org.junit.Assert;
+import org.junit.Test;
 
-public class TestBytesToBytesMapOnHeap extends AbstractTestBytesToBytesMap {
+import org.apache.spark.unsafe.memory.MemoryBlock;
 
-  @Override
-  protected MemoryAllocator getMemoryAllocator() {
-    return MemoryAllocator.HEAP;
+public class LongArraySuite {
+
+  @Test
+  public void basicTest() {
+    long[] bytes = new long[2];
+    LongArray arr = new LongArray(MemoryBlock.fromLongArray(bytes));
+    arr.set(0, 1L);
+    arr.set(1, 2L);
+    arr.set(1, 3L);
+    Assert.assertEquals(2, arr.size());
+    Assert.assertEquals(1L, arr.get(0));
+    Assert.assertEquals(3L, arr.get(1));
   }
-
 }
