+++ unsafe/src/test/java/org/apache/spark/unsafe/map/BytesToBytesMapOnHeapSuite.java
@@ -15,24 +15,15 @@
  * limitations under the License.
  */
 
-package org.apache.spark.unsafe.array;
+package org.apache.spark.unsafe.map;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.apache.spark.unsafe.memory.MemoryAllocator;
 
-import org.apache.spark.unsafe.memory.MemoryBlock;
+public class BytesToBytesMapOnHeapSuite extends AbstractBytesToBytesMapSuite {
 
-public class TestLongArray {
-
-  @Test
-  public void basicTest() {
-    long[] bytes = new long[2];
-    LongArray arr = new LongArray(MemoryBlock.fromLongArray(bytes));
-    arr.set(0, 1L);
-    arr.set(1, 2L);
-    arr.set(1, 3L);
-    Assert.assertEquals(2, arr.size());
-    Assert.assertEquals(1L, arr.get(0));
-    Assert.assertEquals(3L, arr.get(1));
+  @Override
+  protected MemoryAllocator getMemoryAllocator() {
+    return MemoryAllocator.HEAP;
   }
+
 }
